add placeorder button funtionality current cart item show in nav var use subject to service

+++ src/app/Shopping/cart/cart.component.ts
@@ -79,7 +79,7 @@
 // }
 
 
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { CartService } from './cart.service';
 import { Cart, CartKey, CartProduct } from './cart.model';
 
@@ -87,20 +87,21 @@ import { Cart, CartKey, CartProduct } from './cart.model';
 import { CheckoutService } from '../checkout/checkout.service';
 
 import { Router } from '@angular/router';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-cart',
   templateUrl: './cart.component.html',
   styleUrls: ['./cart.component.css']
 })
-export class CartComponent implements OnInit {
+export class CartComponent implements OnInit,OnDestroy {
   cartItems: CartProduct[];
   deleteClicked: boolean = false;
   deleteCartId: number = null;  
   hasData = false;
   isAuthenticate = true;
   userId: number = 3;
-
+  cartChangesSubscription: Subscription;
   cartKey: CartKey = {
     name: 'cart',
     id: this.userId
@@ -119,6 +120,11 @@ export class CartComponent implements OnInit {
       this.cartService.setCartKey('cart', this.userId);
       this.hasData = this.cartService.isDataInLocalStorage();
       this.cartService.saveDataInCart(this.hasData, this.userId);
+      this.cartChangesSubscription = this.cartService.changeOnCart.subscribe(
+        () => {
+          this.cartItems = this.cartService.getCartItems();
+        }
+      )
       this.cartItems = this.cartService.getCartItems();
     }
   }
@@ -150,6 +156,7 @@ export class CartComponent implements OnInit {
     if (status === 'close') {
       this.deleteClicked = false;
     } else {
+      if(this.deleteCartId!==null)
       this.onDeleteCart(this.deleteCartId);
       this.deleteClicked = false;
       this.deleteCartId = null;
@@ -172,6 +179,9 @@ export class CartComponent implements OnInit {
   onSaveCheckout(id:number) {
     let index = this.cartItems.findIndex(cart => cart.productId === id);
     this.cartItems[index].saveForCheckout = !this.cartItems[index].saveForCheckout;
+  }
+  ngOnDestroy(): void {
+    this.cartChangesSubscription.unsubscribe();
   }
 
 }
